Import fix

+++ src/main/java/com/example/Warehouse/repository/jpa/ProductComponentEntityJpaRepository.java
@@ -2,5 +2,5 @@ package com.example.Warehouse.repository.jpa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductComponentEntityJpaRepository extends JpaRepository<ProductComponentEntity, Long> {
+public interface ProductComponentEntityJpaRepository extends JpaRepository<ProductComponentEntity, Integer> {
 }
